requirements.txt
streamlit
supabase
python-dotenv
PyPDF2
openai
json